+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2021-11-27
+
+### Added
+
+-   Bundle optimizations
+-   Styles for notifications
+
 ## [1.5.0] - 2021-11-27
 
 ### Added
@@ -45,7 +52,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Basic features
 
-[Unreleased]: https://github.com/baarsa/opera-guide-typegraphql/compare/1.5.0...HEAD
+[Unreleased]: https://github.com/baarsa/opera-guide-typegraphql/compare/1.6.0...HEAD
+
+[1.6.0]: https://github.com/baarsa/opera-guide-typegraphql/compare/1.5.0...1.6.0
 
 [1.5.0]: https://github.com/baarsa/opera-guide-typegraphql/compare/1.4.0...1.5.0